Update .gitignore templates (#2772)
The new templates are similar to the ones generated by
`mix new`, which contains more detailed descriptions and
more files/directories ignored.

Besides, it appends slashes to the name of directories to
differentiate them from regular files. This change was discussed
on the Elixir core mailing list, and later implemented on the
.gitignore templates of `mix new`:

https://groups.google.com/forum/#!topic/elixir-lang-core/1uf4pC5Iyvw
https://github.com/elixir-lang/elixir/pull/6091

In addition, it includes in the list of ignored files the package
tarballs generated with `mix hex.build`. These files are also
included in the .gitignore templates of `mix new`. This change was
introduced in https://github.com/elixir-lang/elixir/pull/6494.