Replace extract-text-webpack-plugin with mini-css-extract-plugin (#2831)
According to the TL;DR of the comment below, mini-css-extract-plugin is
replacing extract-text-webpack-plugin which will be deprecated. Since
this does not have a built in optimizer for CSS, a separate plugin is
used. However overriding the optimizer means that UglifyJS must be
included manually (previously this was a transitive dependency of
webpack which was used automatically in production mode)

https://github.com/webpack-contrib/extract-text-webpack-plugin/issues/749#issuecomment-374549467

This closes #2818